enablr/enablr/packages/api/register/requirements.txt
aws-lambda-powertools[tracer,parser]
pyjwt
cryptography